+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes for Craft Nitro
 
+## Unreleased
+
+### Fixed
+- Fixed a bug where empty config files would prevent commands from running [#256](https://github.com/craftcms/nitro/issues/256).
+
 ## 2.0.0-beta.7 - 2021-02-25
 
 ### Changed