+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 - Fixed a bug where empty config files would prevent commands from running [#256](https://github.com/craftcms/nitro/issues/256).
+- The `trust` command now properly formats output for WSL users [#257](https://github.com/craftcms/nitro/issues/257).
 
 ## 2.0.0-beta.7 - 2021-02-25